# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple server.",
        "title": "osu!lazer API",
        "contact": {
            "name": "RL GitHub",
            "url": "https://github.com/rl-os/api"
        },
        "license": {},
        "version": "2.0"
    },
    "host": "localhost:2400",
    "basePath": "/",
    "paths": {
...
        "/api/v2/oauth/client": {
            "post": {
                "security": [
                    {
                        "OAuth2": []
                    }
                ],
                "tags": [
                    "OAuth"
                ],
                "summary": "Create new oauth client",
                "parameters": [
                    {
                        "description": "JSON payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateOAuthClient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.OAuthClient"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ResponseFormat"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.OAuthClient": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "redirect": {
                    "type": "string"
                },
                "revoked": {
                    "type": "boolean"
                },
                "secret": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "errors.ResponseFormat": {
            "type": "object",
            "properties": {
                "error_description": {
                    "type": "string"
                },
                "error_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "request.CreateOAuthClient": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "redirect": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "OAuth2": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/oauth/token",
            "scopes": {
                "*": " Grants all access",
                "admin.*": " Grants access as admin",
                "sys.*": " Grants access as system user (for example chatbot, worker and etc)",
                "user.*": " Grants access as user (without system and admin api)"
            }
        }
    }
}